+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Bug Fixes
 ::
 
    * Fixing werkzeug 3 deprecated version import. Import is replaced by new style version check with importlib (#573) [Ryu-CZ]
+   * Fixing flask 3.0+ compatibility of `ModuleNotFoundError: No module named 'flask.scaffold'` Import error. (#567) [Ryu-CZ]
 
 
 .. _section-1.2.0: